Modifications to map structure
WIP, doesn't work. Implementing map class
