+++ src/data/tile0.txt
@@ -1,3 +1,3 @@
 src/txtr/0.png
 false
-null
+null ~
