+++ src/data/tile1.txt
@@ -1,3 +1,3 @@
 src/txtr/1.png
 true
-null
+null ~
