+++ src/data/tile2.txt
@@ -1,3 +1,3 @@
 src/txtr/2.png
 false
-This tile is different from the rest... -
+This tile is different from the rest... ~
